Fix Mono and GDNative builds after changes to ObjectID
Issues caused by cf8c679a23b21d6c6f29cba6a54eaa2eed88bf92.

The Mono change is actually a bugfix (used the int instead of ObjectID
by mistake).

The GDNative change is a temporary revert until a more exhaustive approach
is taken to make 'godot_int' 64-bit, is confirmed wanted by GDNative users.

+++ modules/gdnative/include/gdnative/gdnative.h
@@ -127,7 +127,7 @@ typedef bool godot_bool;
 
 /////// int
 
-typedef int64_t godot_int;
+typedef int godot_int;
 
 /////// real
